feat: packed reader and writer new api (#172)
main feat for the patch:
* Column groups and file paths are explicitly defined by users. The
packed writer will not split column groups by memory. It follows the
user-defined column groups to split arrow records and write into the
corresponding file path.
* Change factory arrow filesystem to singleton arrow file system to
support Milvus init core.
* Separate arrow fs config from storage config.
related: #171

---------

+++ cpp/include/milvus-storage/packed/column_group.h
@@ -61,7 +61,7 @@ class ColumnGroup {
 
   Status Clear();
 
-  int GetTotalRows() const { return total_rows_; }
+  int64_t GetTotalRows() const { return total_rows_; }
 
   private:
   GroupId group_id_;
@@ -69,7 +69,7 @@ class ColumnGroup {
   std::vector<std::shared_ptr<arrow::RecordBatch>> batches_;
   size_t memory_usage_;
   std::vector<int> origin_column_indices_;
-  int total_rows_;
+  int64_t total_rows_;
 };
 
 struct ColumnGroupState {
